Update file strucutre
