+++ sheerwater_benchmarking/models/__init__.py
@@ -1,7 +1,5 @@
 """Forecasting models for the Sheerwater benchmarking project."""
-# from .salient import salient_forecast
+from .salient import salient_era5_raw
 
-# # Use __all__ to define what is part of the public API.
-# __all__ = [
-#     salient_forecast,
-# ]
+# Use __all__ to define what is part of the public API.
+__all__ = []
