+++ sheerwater_benchmarking/reanalysis/__init__.py
@@ -1,10 +1,9 @@
 """Data functions for all parts of the data pipeline."""
-from .era5 import era5, era5_rolled, era5_agg, salient_era5
+from .era5 import era5, era5_rolled, era5_agg
 
 # Use __all__ to define what is part of the public API.
 __all__ = [
     era5,
     era5_rolled,
     era5_agg,
-    salient_era5,
 ]
