+++ sheerwater_benchmarking/reanalysis/era5.py
@@ -4,15 +4,11 @@ import cdsapi
 import os
 import xarray as xr
 import dateparser
-import numpy as np
 
-import salientsdk as sk
-
-from sheerwater_benchmarking.utils import dask_remote, cacheable
-from sheerwater_benchmarking.utils.secrets import cdsapi_secret, salient_auth
-from sheerwater_benchmarking.utils.general_utils import get_grid, get_variable
-from sheerwater_benchmarking.utils.data_utils import apply_mask, roll_and_agg, regrid
-from sheerwater_benchmarking.utils.model_utils import get_salient_loc
+from sheerwater_benchmarking.utils import (dask_remote, cacheable,
+                                           cdsapi_secret,
+                                           get_grid, get_variable,
+                                           apply_mask, roll_and_agg, regrid)
 from sheerwater_benchmarking.masks import land_sea_mask
 
 
@@ -210,78 +206,4 @@ def era5_agg(start_time, end_time, variable, grid="global1_5", agg=14, mask="lsm
 
     ds = apply_mask(ds, mask_ds, variable, val=0.0, rename_dict={"mask": variable})
 
-    return ds
-
-
-@salient_auth
-@cacheable(data_type='array',
-           timeseries='time',
-           cache_args=['variable', 'grid'],
-           chunking={"lat": 292, "lon": 396, "time": 300},
-           auto_rechunk=True)
-def salient_era5_raw(start_time, end_time, variable, grid="africa0_25", verbose=False):
-    """Fetches ground truth data from Salient's SDK and applies aggregation and masking .
-
-    Args:
-        start_time (str): The start date to fetch data for.
-        end_time (str): The end date to fetch.
-        variable (str): The weather variable to fetch.
-        grid (str): The grid resolution to fetch the data at. One of:
-            - africa0_25: 0.25 degree African grid
-        verbose (bool): Whether to print verbose output.
-    """
-    # Fetch the data from Salient
-    loc = get_salient_loc(grid)
-    var_name = {'tmp2m': 'temp', 'precip': 'precip'}[variable]
-
-    # Fetch and load the data
-    data = sk.data_timeseries(
-        loc=loc,
-        variable=var_name,
-        field="vals",
-        start=np.datetime64(start_time),
-        end=np.datetime64(end_time),
-        frequency="daily",
-        verbose=verbose,
-        force=True,
-    )
-    return xr.load_dataset(data)
-
-
-@cacheable(data_type='array',
-           timeseries='time',
-           cache_args=['variable', 'grid', 'agg', 'mask'],
-           chunking={"lat": 292, "lon": 396, "time": 500},
-           auto_rechunk=False)
-def salient_era5(start_time, end_time, variable, grid="africa0_25",
-                 agg=14, mask="lsm", verbose=False):
-    """Fetches ground truth data from Salient's SDK and applies aggregation and masking .
-
-    Args:
-        start_time (str): The start date to fetch data for.
-        end_time (str): The end date to fetch.
-        variable (str): The weather variable to fetch.
-        grid (str): The grid resolution to fetch the data at. One of:
-            - global1_5: 1.5 degree global grid
-        agg (str): The aggregation period to use, in days
-        mask (str): The mask to apply to the data. One of:
-            - lsm: Land-sea mask
-            - None: No mask
-        verbose (bool): Whether to print verbose output.
-    """
-    # Get raw salient data
-    ds = salient_era5_raw(start_time, end_time, variable, grid=grid, verbose=verbose)
-
-    agg_fn = "sum" if variable == "precip" else "mean"
-    ds = roll_and_agg(ds, agg=agg, agg_col="time", agg_fn=agg_fn)
-
-    if mask == "lsm":
-        # Select variables and apply mask
-        mask_ds = land_sea_mask(grid=grid).compute()
-    elif mask is None:
-        mask_ds = None
-    else:
-        raise NotImplementedError("Only land-sea or None mask is implemented.")
-
-    ds = apply_mask(ds, mask_ds, variable, val=0.0, rename_dict={"mask": variable})
     return ds
