+++ sheerwater_benchmarking/utils/__init__.py
@@ -1,13 +1,17 @@
 """Utility functions for benchmarking."""
-from .caching import cacheable
+from .caching_utils import cacheable
 from .remote import dask_remote
 from .secrets import cdsapi_secret, ecmwf_secret, salient_secret, salient_auth
 from .data_utils import apply_mask, roll_and_agg, regrid
-from .general_utils import get_grid, get_variable
+from .general_utils import (load_netcdf, write_zarr, load_zarr,
+                            is_valid_forecast_date,
+                            get_dates, get_variable, get_grid)
+
 
 # Use __all__ to define what is part of the public API.
 __all__ = [
     cacheable,
+    dask_remote,
     cdsapi_secret,
     ecmwf_secret,
     salient_secret,
@@ -15,7 +19,11 @@ __all__ = [
     apply_mask,
     roll_and_agg,
     regrid,
-    get_grid,
+    load_netcdf,
+    write_zarr,
+    load_zarr,
+    is_valid_forecast_date,
+    get_dates,
     get_variable,
-    dask_remote,
+    get_grid,
 ]
